==> 2-Conference Join.ipynb <==
# Join Conference Information

import pandas as pd

df = pd.read_csv('FINAL_DF.csv')
df

conf = pd.read_csv('CONF.CSV').drop_duplicates()
conf

def check_names(dataset1, dataset2):
    bad_names = []
    for w in dataset1["Team"]:
        if w not in dataset2["School"].values and w not in dataset2["Opponent"].values and w not in bad_names:
            bad_names.append(w)
    print(len(bad_names))
    print(bad_names)

check_names(conf, df)
df_fix_names = conf
df_fix_names = df_fix_names.replace('Ohio State', 'Ohio')
df_fix_names = df_fix_names.replace('LSU', 'Louisiana State')
df_fix_names = df_fix_names.replace('Ole Miss', 'Mississippi')
df_fix_names = df_fix_names.replace('Pitt', 'Pittsburgh')
df_fix_names = df_fix_names.replace('SMU', 'Southern Methodist')
df_fix_names = df_fix_names.replace('UAB', 'Alabama-Birmingham')
df_fix_names = df_fix_names.replace('UCF', 'Central Florida')
df_fix_names = df_fix_names.replace('UCF', 'Central Florida')
df_fix_names = df_fix_names.replace('UNLV', 'Nevada-Las Vegas')
df_fix_names = df_fix_names.replace('USC', 'Southern California')
df_fix_names = df_fix_names.replace('UTEP', 'Texas-El Paso')
df_fix_names = df_fix_names.replace('UTSA', 'TExas-San Antonio')
df_fix_names = df_fix_names.replace("Mississippi St", "Mississippi State")
df_fix_names = df_fix_names.replace("LSU", "Louisiana State")
df_fix_names = df_fix_names.replace("Miami Florida", "Miami (FL)")
df_fix_names = df_fix_names.replace("NC State", "North Carolina State")
df_fix_names = df_fix_names.replace("Florida Intl", "Florida International")
df_fix_names = df_fix_names.replace("USC", "Southern California")
df_fix_names = df_fix_names.replace("TCU", "Texas Christian")
df_fix_names = df_fix_names.replace("UL Lafayette", "Lafayette")
df_fix_names = df_fix_names.replace("Appalachian St", "Appalachian State")
df_fix_names = df_fix_names.replace("SMU", "Southern Methodist")
df_fix_names = df_fix_names.replace("Bowling Green", "Bowling Green State")
df_fix_names = df_fix_names.replace("BYU", "Brigham Young")
df_fix_names = df_fix_names.replace("Miami Ohio", "Miami (OH)")
df_fix_names = df_fix_names.replace("Southern Miss", "Southern Mississippi")
df_fix_names = df_fix_names.replace("UNLV", "Nevada-Las Vegas")
df_fix_names = df_fix_names.replace("UTEP", "Texas-El Paso")
df_fix_names = df_fix_names.replace("UTSA", "Texas-San Antonio")
df_fix_names = df_fix_names.replace("Middle Tenn St", "Middle Tennessee State")
df_fix_names = df_fix_names.replace("Kent", "Kent State")
df_fix_names = df_fix_names.replace("Charlotte U", "Charlotte")
df_fix_names = df_fix_names.replace("UAB", "Alabama-Birmingham")
df_fix_names = df_fix_names.replace("NW Missouri State", "Missouri State")
df_fix_names = df_fix_names.replace("Mary Hardin-Baylor", "Baylor")
df_fix_names = df_fix_names.replace("Texas A&M Commerce", "Texas A&M")
df_fix_names = df_fix_names.replace("McNeese St", "McNeese State")
df_fix_names = df_fix_names.replace("Tennessee Chat", "Chattanooga")
df_fix_names = df_fix_names.replace('FIU', 'Florida International')
df_fix_names = df_fix_names.replace('Louisiana', 'Louisiana State')
df_fix_names = df_fix_names.replace('Louisiana–Monroe', 'Louisiana-Monroe')
df_fix_names = df_fix_names.replace('Middle Tennessee', 'Middle Tennessee State')
df_fix_names = df_fix_names.replace('UConn', 'Connecticut')
df_fix_names = df_fix_names.replace('UMass', 'Massachusetts')
df_fix_names = df_fix_names.replace('TExas-San Antonio', 'Texas-San Antonio')
check_names(df_fix_names, df)
conf = df_fix_names
conf[(conf["Team"] == "Oklahoma State") | (conf["Team"] == "Central Michigan") ]
print(len(conf.drop_duplicates()))

school_merged = pd.merge(left=df, right=conf, how='left', left_on=['School'], right_on=['Team'])

opp_merged = pd.merge(left=df, right=conf, how='left', left_on=['Opponent'], right_on=['Team'])
opp_merged = opp_merged.rename(columns={
    "Team": "OppTeam",
    "Nickname": "OppNickname",
    "City": "OppCity",
    "State": "OppState",
    "Enrollment": "OppEnrollment",
    "Conference": "OppConference"
})[["School", "Opponent", "Date", "OppNickname", "OppCity", "OppState", "OppEnrollment", "OppConference"]]
merge_on = ['School', 'Opponent', 'Date']

merged_conferences = pd.merge(left=school_merged, right=opp_merged, how='inner', left_on=merge_on, right_on=merge_on).drop_duplicates().dropna()

merged_conferences.to_csv('FINAL_DF_CONFERENCES.csv', index=False)
