==> 3-Diff Stats.ipynb <==
# Convert all season stats to differences

import pandas as pd

df = pd.read_csv('FINAL_DF_CONFERENCES.csv')
df

df.columns

### Make Diff columns

for c in df.columns:
    print("df[\"" + c + "Diff\"] = df[\"" + c + "\"] - df[\"Opp" + c + "\"]")

df["PassCmpDiff"] = df["PassCmp"] - df["OppPassCmp"]
df["PassAttDiff"] = df["PassAtt"] - df["OppPassAtt"]
df["PassPctDiff"] = df["PassPct"] - df["OppPassPct"]
df["PassYdsDiff"] = df["PassYds"] - df["OppPassYds"]
df["PassTDDiff"] = df["PassTD"] - df["OppPassTD"]
df["RushAttDiff"] = df["RushAtt"] - df["OppRushAtt"]
df["RushYdsDiff"] = df["RushYds"] - df["OppRushYds"]
df["RushAvgDiff"] = df["RushAvg"] - df["OppRushAvg"]
df["RushTDDiff"] = df["RushTD"] - df["OppRushTD"]
df["XPMDiff"] = df["XPM"] - df["OppXPM"]
df["XPADiff"] = df["XPA"] - df["OppXPA"]
df["XPPercentDiff"] = df["XPPercent"] - df["OppXPPercent"]
df["FGMDiff"] = df["FGM"] - df["OppFGM"]
df["FGADiff"] = df["FGA"] - df["OppFGA"]
df["FGPercentDiff"] = df["FGPercent"] - df["OppFGPercent"]
df["KickPtsDiff"] = df["KickPts"] - df["OppKickPts"]
df["FumDiff"] = df["Fum"] - df["OppFum"]
df["IntDiff"] = df["Int"] - df["OppInt"]
df["TotalTODiff"] = df["TotalTO"] - df["OppTotalTO"]
df["Enrollment"] = df["Enrollment"].str.replace(',', '')
df["OppEnrollment"] = df["OppEnrollment"].str.replace(',', '')
df["Enrollment"] = pd.to_numeric(df["Enrollment"])
df["OppEnrollment"] = pd.to_numeric(df["OppEnrollment"])
df["EnrollmentDiff"] = df["Enrollment"] - df["OppEnrollment"]

### Delete Old columns

cols = df.columns
for c in cols:
    print("del df[\"" + c + "\"]")

del df["PassCmp"]
del df["PassAtt"]
del df["PassPct"]
del df["PassYds"]
del df["PassTD"]
del df["RushAtt"]
del df["RushYds"]
del df["RushAvg"]
del df["RushTD"]
del df["XPM"]
del df["XPA"]
del df["XPPercent"]
del df["FGM"]
del df["FGA"]
del df["FGPercent"]
del df["KickPts"]
del df["Fum"]
del df["Int"]
del df["TotalTO"]
del df["OppPassCmp"]
del df["OppPassAtt"]
del df["OppPassPct"]
del df["OppPassYds"]
del df["OppPassTD"]
del df["OppRushAtt"]
del df["OppRushYds"]
del df["OppRushAvg"]
del df["OppRushTD"]
del df["OppXPM"]
del df["OppXPA"]
del df["OppXPPercent"]
del df["OppFGM"]
del df["OppFGA"]
del df["OppFGPercent"]
del df["OppKickPts"]
del df["OppFum"]
del df["OppInt"]
del df["OppTotalTO"]
del df["Enrollment"]
del df["OppEnrollment"]

df.columns

# Cleaning

df = df[abs(df["Spread"]) < 54]

# Export

df.to_csv('DF_FINAL_CONFERENCES_DIFFS.csv', index=False)
