==> 4-Cleanup.ipynb <==
# Remove unnecessary columns

import pandas as pd
import numpy as np

df = pd.read_csv('DF_FINAL_CONFERENCES_DIFFS.csv')
len(df.columns)

cols = df.columns
for c in cols:
    print("\"" + c + "\",")

new_df = df.drop(columns=[
    "School",
    "Opponent",
    "Date",
    "Season",
    "NeutralLocation",
    "Season.Type",
    "Score",
    "OppScore",
    "SeasonRanking",
    "OppSeasonRanking",
    "Team",
    "Nickname",
    "City",
    "State",
    "OppNickname",
    "OppCity",
    "OppState",
    "FumDiff",
    "IntDiff",
    "PassCmpDiff",
    "PassAttDiff",
    "PassPctDiff",
    "PassTDDiff",
    "RushAvgDiff",
    "RushTDDiff",
    "XPPercentDiff",
    "FGMDiff",
    "FGADiff",
    "FGPercentDiff",
    "KickPtsDiff",
])

print(new_df.columns)
len(new_df.columns)

new_df['Outcome'] = np.where(new_df['Outcome']=='W', 1, 0)
new_df

# Final Exports

total = new_df.drop(columns=['Spread', 'Outcome'])
result = new_df.drop(columns=['Total', 'Spread'])
spread = new_df.drop(columns=['Total', 'Outcome'])
total.to_csv('df_total.csv', index=False)
result.to_csv('df_result.csv', index=False)
spread.to_csv('df_spread.csv', index=False)
new_df.to_csv('df_reduced_columns.csv', index=False)

# Printing Variables

new_df.columns
for i in range(len(new_df.columns)):
    print(new_df.columns[i])

cols = new_df.columns
pd.DataFrame(rows=

["AtHome", 0],
["Total", 0],
["Outcome", 0],
["Spread", 0],
["GameNumber", 0],
["RankDiff", 0],
["Conference", 0],
["OppConference", 0],
["PassYdsDiff", 0],
["RushAttDiff", 0],
["RushYdsDiff", 0],
["XPMDiff", 0],
["XPADiff", 0],
["TotalTODiff", 0],
["EnrollmentDiff", 0]
)
